Make the library fully device-agnostic (CUDA/CPU/MPS)
Fix the two remaining CUDA blockers and two dtype hardcodes found in a
device audit:

- inference.sample: draw the innovations on the generator's device and
  move them to the model's device, so sampling works on CUDA/MPS models
  and a CPU generator yields device-independent, bit-reproducible
  streams (verified: identical CPU-vs-MPS draws for the same seed).
- reliability: create the full-rank-branch null-projection mask on the
  CRB's device (torch.where mixed CPU/GPU tensors before).
- identifiability.fisher_metric / training.marginal_likelihood: let the
  accumulators follow the computation's real dtype instead of a
  hardcoded float64 (required on MPS, which lacks float64).

Verified: 179 CPU tests pass unchanged; MPS float32 smoke covers
k_full, marginal, sample, Cholesky-path MI, parameter_fisher,
identifiability_report, crb_report, marginal_likelihood, and em_fit
(linalg.eigh via the official PYTORCH_ENABLE_MPS_FALLBACK). README's
GPU/MPS limitation note updated to match.

+++ gaussian_bn/identifiability.py
@@ -84,7 +84,7 @@ def fisher_metric(
     # with the field constant c = 1/2 (real) or 1 (circular complex).
     c = 1.0 if K0.is_complex() else 0.5
     KinvdK = [solve_psd(K0, dKa, jitter=jitter) for dKa in dK]
-    G = torch.zeros((q, q), dtype=torch.float64, device=K0.device)
+    G = torch.zeros((q, q), dtype=K0.real.dtype, device=K0.device)
     for a in range(q):
         for b in range(q):
             G[a, b] = c * torch.trace(KinvdK[a] @ KinvdK[b]).real
